# service/service/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "SB API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:3000",
    "basePath": "/",
    "paths": {
...
        "/messenger/dialogs": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Messenger"
                ],
                "summary": "Get user's dialogs and requests",
                "operationId": "getDialogs",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/handlers.Request"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    }
                }
            }
        },
        "/messenger/messages": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Messenger"
                ],
                "summary": "Get dialog",
                "operationId": "getMessages",
                "parameters": [
                    {
                        "description": "filter to display messages (id_users have to contain to ID, if created_after not defined - considered as current moment)",
                        "name": "Body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.MessagesFilter"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/handlers.Message"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    }
                }
            }
        },
        "/messenger/request/reply": {
            "put": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Messenger"
                ],
                "summary": "Accept or decline friendship or joining a group training request",
                "operationId": "replyToRequest",
                "parameters": [
                    {
                        "description": "The body of a request - status : declined/accepted",
                        "name": "Body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.Request"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    }
                }
            }
        },
        "/messenger/request/seen": {
            "put": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Messenger"
                ],
                "summary": "Mark response to a request as seen",
                "operationId": "seenRequestReply",
                "parameters": [
                    {
                        "description": "The body of a request - seen : true",
                        "name": "Body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.Request"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/user/requests": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Messenger"
                ],
                "summary": "Send friendship or joining a group training request",
                "operationId": "sendRequest",
                "parameters": [
                    {
                        "description": "The body of a request",
                        "name": "Body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.Request"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "ErrorResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "description": "Message about the error",
                    "type": "string",
                    "example": "Operation failed"
                }
            }
        },
...
        "handlers.Message": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string",
                    "example": "Hello!"
                },
                "created_at": {
                    "type": "string",
                    "example": "2021-11-29T00:16:01.367333+03:00"
                },
                "id_from": {
                    "type": "integer",
                    "example": 321
                },
                "id_to": {
                    "type": "integer",
                    "example": 123
                },
                "type": {
                    "type": "string",
                    "example": "personal"
                }
            }
        },
        "handlers.MessagesFilter": {
            "type": "object",
            "properties": {
                "created_after": {
                    "type": "string"
                },
                "id_users": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                }
            }
        },
        "handlers.Request": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string",
                    "example": "2021-11-29T00:16:01.367333+03:00"
                },
                "id_from": {
                    "type": "integer",
                    "example": 321
                },
                "id_to": {
                    "type": "integer",
                    "example": 123
                },
                "seen": {
                    "type": "boolean",
                    "example": false
                },
                "status": {
                    "type": "string",
                    "example": "declined"
                },
                "type": {
                    "type": "string",
                    "example": "personal"
                }
            }
        },
...
    }
}